requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
google-generativeai>=0.8.3
gunicorn==21.2.0